Changed float to double in exercise2.h

+++ week6/Assignment_6.4.1/include/exercise2.h
@@ -11,9 +11,9 @@ namespace vector_namespace
     class Vector3D
     {
     public:
-        float x{};
-        float y{};
-        float z{};
+        double x{};
+        double y{};
+        double z{};
     };
 
     // operator overloads
